% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_random_buysell_engine.R
\name{apply_random_buysell_engine}
\alias{apply_random_buysell_engine}
\title{Apply a Random Buy/Sell Engine to Stock Data}
\usage{
apply_random_buysell_engine(stock_data, turnover_per_year)
}
\arguments{
\item{stock_data}{An xts object containing the stock's data. Must have a 'Close' column.}

\item{turnover_per_year}{Numeric, representing the desired turnover rate per year.}
}
\value{
An xts object that includes the original stock data augmented with 'Signal', 'Execution', and 'Hold' columns.
        'Signal' represents the generated buy/sell signals, 'Execution' reflects the execution of these signals,
        and 'Hold' indicates whether the position is held (1) or not (0) after the execution of a signal.
}
\description{
This function generates a random buy/sell signal for a given stock data based on the specified turnover rate per year.
It ensures an alternating pattern between buy (1) and sell (-1) signals, starting with a buy signal.
The function creates a series of buy and sell signals spread randomly over the trading period,
maintaining an average holding period corresponding to the specified turnover rate.
}
\examples{
# Generate sample stock data
stock_data <- generate_continuous_price(0.0005, 0.0124, n = 1000)

# Apply random buy/sell engine with a specified annual turnover rate
buysell_engine <- apply_random_buysell_engine(stock_data, turnover_per_year = 12)

# View the resulting buy/sell signals
print(buysell_engine)

}
