% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_volatility.R
\name{comp_volatility}
\alias{comp_volatility}
\title{Compute Exponentially Weighted Moving Volatility}
\usage{
comp_volatility(stock_data, window = 36)
}
\arguments{
\item{stock_data}{An xts object containing the stock's closing prices.}

\item{window}{An integer specifying the look-back window for the exponential decay factor.
Default is set to 36.}
}
\value{
An xts object containing the computed price volatility values for each date in the original dataset.
}
\description{
This function computes the exponentially weighted moving volatility of a given stock
based on its closing prices. The exponential decay factor is utilized to give more weight
to recent observations.
}
\examples{
stock_data = xts(runif(100, 100, 200), Sys.Date() + 1:100)
comp_volatility(stock_data)
comp_volatility(stock_data, window = 50)
}
