% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_discrete_price.R
\name{generate_discrete_price}
\alias{generate_discrete_price}
\title{Generate Discrete Price Series Based on a Payoff and Win Chance}
\usage{
generate_discrete_price(payoff, win_chance, n)
}
\arguments{
\item{payoff}{Numeric, the amount won in each winning case.}

\item{win_chance}{Numeric, the probability of winning between 0 and 1.}

\item{n}{Integer, the number of time steps to simulate.}
}
\value{
An xts object containing the equity values over the n time steps.
}
\description{
This function generates a discrete series of equity prices based on given
payoff and winning chance parameters. The function starts with an initial
equity of 100 and simulates each discrete step based on the provided
win_chance and payoff parameters.
}
\examples{
generate_discrete_price(payoff = 2, win_chance = 0.5, n = 100)
}
