% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trendy_price.R
\name{generate_noise}
\alias{generate_noise}
\title{Generate Gaussian Noise Sequence}
\usage{
generate_noise(Nlength, stdev)
}
\arguments{
\item{Nlength}{The length of the time series.}

\item{stdev}{The standard deviation of the Gaussian noise.}
}
\value{
A vector of Gaussian noise.
}
\description{
This function creates a sequence of Gaussian noise data points.
}
