% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trendy_price.R
\name{generate_siney_trends}
\alias{generate_siney_trends}
\title{Create Sine-Wave Trend Data}
\usage{
generate_siney_trends(Nlength, Tlength, Xamplitude)
}
\arguments{
\item{Nlength}{The total number of data points.}

\item{Tlength}{The period of one sine wave cycle.}

\item{Xamplitude}{The amplitude of the sine wave.}
}
\value{
A vector containing the sine-wave trend.
}
\description{
This function generates a sine-wave trend for a time series.
}
