% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trendy_price.R
\name{generate_trends}
\alias{generate_trends}
\title{Generate Linear Trend Data}
\usage{
generate_trends(Nlength, Tlength, Xamplitude)
}
\arguments{
\item{Nlength}{The total number of data points.}

\item{Tlength}{The period for one complete up and down cycle.}

\item{Xamplitude}{The amplitude of the linear trend.}
}
\value{
A vector containing the linear trend.
}
\description{
This function generates a linear trend for a time series.
}
