% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trendy_price.R
\name{generate_trendy_price}
\alias{generate_trendy_price}
\title{Create Synthetic Price Series}
\usage{
generate_trendy_price(Nlength, Tlength, Xamplitude, Volscale, sines = FALSE)
}
\arguments{
\item{Nlength}{The total number of data points.}

\item{Tlength}{The period for the trend.}

\item{Xamplitude}{The amplitude of the trend.}

\item{Volscale}{The volatility scaling factor.}

\item{sines}{If TRUE, uses sine-wave trends. If FALSE, uses linear trends.}
}
\value{
An xts object containing the synthetic price series.
}
\description{
This function generates a synthetic price series based on trends and Gaussian noise.
}
