% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_setsmart.R
\name{import_data_setsmart}
\alias{import_data_setsmart}
\title{Import End-of-Day Stock Data from SETSMART}
\usage{
import_data_setsmart(api_key, symbol, startDate, endDate)
}
\arguments{
\item{api_key}{string The API key for authenticating with the SETSMART API.}

\item{symbol}{string The stock symbol for which data is being requested.}

\item{startDate}{string The start date for the data retrieval in 'YYYY-MM-DD' format.}

\item{endDate}{string The end date for the data retrieval in 'YYYY-MM-DD' format.}
}
\value{
An 'xts' object containing the columns: Date, Open, High, Low, Close.
        Each row represents a day's trading data for the specified symbol and date range.
        In case of an error or unavailable data, a warning is issued, and NULL is returned.
}
\description{
Retrieves End-of-Day stock data from the SETSMART API for a specified symbol over a given date range.
The function returns data in an 'xts' format with columns for Open, High, Low, and Close prices.
}
\examples{
\dontrun{
api_key <- "your_api_key"
stock_data <- import_data_setsmart(api_key, "TGE", "2024-01-04", "2024-01-15")
plot(stock_data)
}
}
