% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_stooq.R
\name{import_data_stooq}
\alias{import_data_stooq}
\title{Import Data from Stooq}
\usage{
import_data_stooq(symbol, start_date, end_date)
}
\arguments{
\item{symbol}{The ticker symbol of the stock.}

\item{start_date}{The start date for the data in 'YYYY-MM-DD' format.}

\item{end_date}{The end date for the data in 'YYYY-MM-DD' format.}
}
\value{
An xts object containing the stock data.
}
\description{
This function imports historical stock data from Stooq and converts it to an xts object.
It formats the start and end dates to the required format and constructs the URL to download the data.
}
\examples{
\dontrun{
apple_stock <- import_data_stooq("AAPL.US", "2020-01-01", "2021-01-01")
}

}
