% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_ema_crossover_buysell_engine.R
\name{apply_ema_crossover_buysell_engine}
\alias{apply_ema_crossover_buysell_engine}
\title{Apply EMA Crossover Buy/Sell Engine}
\usage{
apply_ema_crossover_buysell_engine(
  stock_data,
  ema_fast_param,
  ema_slow_param,
  lag = 1
)
}
\arguments{
\item{stock_data}{A data frame or xts object containing stock data, must include a 'Close' column.}

\item{ema_fast_param}{Integer specifying the period for the fast EMA.}

\item{ema_slow_param}{Integer specifying the period for the slow EMA.}

\item{lag}{Integer specifying the lag for Hold state, default is 1.}
}
\value{
An xts object that includes the original stock data, Ema_fast, Ema_slow, Signal, Execution and Hold.
}
\description{
This function applies an Exponential Moving Average (EMA) crossover strategy to a given set of stock data.
It calculates the EMA based on specified parameters for fast and slow EMAs and generates buy/sell signals
and Hold states based on these EMAs. The function returns the stock data augmented with EMA values,
signals, and Hold states.
}
\details{
Signal = {-1, 0 ,1} generated by end of time (t)
Execution = {-1, 0, 1} generated by end of time (t) + lag
Hold = {1, 0} assume hold position from (t+1) + lag to sell signal at time (t)
buy signal(t) then buy at ATO(t+1) then sell signal(t) then sell at ATO(t+1)
}
\examples{
# Example usage:
# stock_data should be an xts object with a 'Close' column
result <- apply_ema_crossover_buysell_engine(stock_data, 12, 26)

}
