% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{checking_data}
\alias{checking_data}
\title{Check Data for Common Issues}
\usage{
checking_data(stock_data)
}
\arguments{
\item{stock_data}{An xts or data frame object containing stock data.}
}
\value{
A matrix with counts of various data issues:
        - 'number_of_NA': Number of missing values (NA) in the data.
        - 'number_of_duplication': Number of duplicated indices in the data.
        - 'number_of_negative': Number of negative values in the data.
}
\description{
This function checks a given stock data set for common data issues such as missing values (NA),
duplicated indices, and negative values. It counts the number of occurrences of these issues.
}
\examples{
# Assuming stock_data is an xts or data frame object
issues <- checking_data(stock_data)
print(issues)

}
