% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_ewma_volatility.R
\name{comp_ewma_volatility}
\alias{comp_ewma_volatility}
\title{Compute Exponentially Weighted Moving Average (EWMA) Volatility}
\usage{
comp_ewma_volatility(stock_data, window = 36)
}
\arguments{
\item{stock_data}{An xts object containing the stock's closing prices.}

\item{window}{An integer specifying the look-back window for the exponential decay factor.
The default is set to 36.}
}
\value{
An xts object containing the computed EWMA volatility values for each date
        in the original dataset, rounded to four decimal places.
}
\description{
This function computes the Exponentially Weighted Moving Average (EWMA) volatility
of a given stock's returns. The function uses a decay factor to assign more weight to
recent observations. The calculation is based on the square of daily returns.
}
\examples{
stock_data = xts(runif(100, 100, 200), Sys.Date() + 1:100)
comp_ewma_volatility(stock_data)
comp_ewma_volatility(stock_data, window = 50)
}
