% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stock_information.R
\name{comp_stock_information}
\alias{comp_stock_information}
\title{Compile Stock Information from Multiple Symbols}
\usage{
comp_stock_information(universe, start_date, end_date)
}
\arguments{
\item{universe}{A vector of stock symbols (tickers) for which the data needs to be fetched.}

\item{start_date}{The start date for the data fetch in "YYYY-MM-DD" format.}

\item{end_date}{The end date for the data fetch in "YYYY-MM-DD" format.}
}
\value{
A list containing two elements:
        - 'daily_return': A data frame (or an xts object) with columns representing the daily returns for each stock symbol.
        - 'close_price': A data frame (or an xts object) with columns representing the closing prices for each stock symbol.
}
\description{
This function gathers stock data for a given list of symbols over a specified date range.
It fetches the daily closing prices and computes the daily returns for each stock symbol.
The function is designed to handle cases where some stock data might not be available or encounters an error during fetching.
Data source: stooq.com
}
\examples{
stocks_universe <- c("AAPL.US", "MSFT.US", "GOOGL.US")
stock_info <- comp_stock_information(stocks_universe, "2020-01-01", "2021-01-01")
head(stock_info$daily_return)
head(stock_info$close_price)

}
