% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stock_data_setsmart.R
\name{download_stock_data_setsmart}
\alias{download_stock_data_setsmart}
\title{Download Stock Data by Symbol from SETSMART API}
\usage{
download_stock_data_setsmart(api_key, symbol, startDate, endDate)
}
\arguments{
\item{api_key}{string The API key for authentication with the SETSMART API.}

\item{symbol}{string The stock symbol for which data is being requested.}

\item{startDate}{Date The start date for the data retrieval in 'YYYY-MM-DD' format.}

\item{endDate}{Date The end date for the data retrieval in 'YYYY-MM-DD' format.}
}
\value{
A dataframe containing the stock data including prices and statistics for the specified symbol and date range.
        In case of an error or unavailable data, a warning is issued and an empty dataframe is returned.
}
\description{
Retrieves end-of-day (EOD) stock price data for a given symbol from the SETSMART API. 
This data includes both original and adjusted prices along with other relevant statistics. 
The data availability period is subject to the SETSMART package's constraints.
}
\examples{
api_key <- "your_api_key"
stock_quote <- download_stock_data_setsmart(api_key, "JTS", "2023-01-02", "2023-01-04")

}
