% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_googlesheet.R
\name{import_data_googlesheet}
\alias{import_data_googlesheet}
\title{Read Stock Data from Google Sheets and Convert to xts Object}
\usage{
import_data_googlesheet(sheet_url, sheet_name)
}
\arguments{
\item{sheet_url}{A string containing the URL of the Google Sheets document.}

\item{sheet_name}{A string specifying the name of the sheet within the Google Sheets document.}
}
\value{
An xts object containing the stock data, with the first column assumed to be dates.
}
\description{
This function reads stock data from a specified Google Sheet and converts it
into an xts object for time series analysis. The function requires the URL of
the Google Sheet and the name of the specific sheet within the document.
}
\examples{
sheet_url <- "https://docs.google.com/spreadsheets/d/..."
stock_data <- read_stock_data_from_sheet(sheet_url, "Sheet1")

}
