% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setsmart_get_database.R
\name{setsmart_get_database}
\alias{setsmart_get_database}
\title{Get End-of-Day Stock Data for Common Stocks from API}
\usage{
setsmart_get_database(api_key, from, to)
}
\arguments{
\item{api_key}{string The API key for authentication with the data source.}

\item{from}{Date The start date for the data retrieval in 'YYYY-MM-DD' format.}

\item{to}{Date The end date for the data retrieval in 'YYYY-MM-DD' format.}
}
\value{
A dataframe with columns: date, symbol, open, high, low, close, aomVolume, aomValue.
        Each row represents a day's trading data for a specific symbol within the given date range.
        If data for a specific date is unavailable or an error occurs, a warning is issued, and an empty dataframe is returned.
}
\description{
Retrieves end-of-day (EOD) stock price data for common stocks ('CS') over a specified date range. 
The data includes open, high, low, and close prices, as well as volume and value of trades.
This function automatically sets the security type to 'CS' and the adjusted price flag to 'Y'.
}
\examples{
api_key <- "your_api_key"
stock_data <- setsmart_get_database(api_key, "2021-01-01", "2021-01-31")

}
